+++ db/migrate/20180705121340_add_tagline_to_service.rb
@@ -1,0 +1,5 @@
+class AddTaglineToService < ActiveRecord::Migration[5.2]
+  def change
+    add_column :services, :tagline, :text, null: false
+  end
+end