+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Possibility to configure development DB using ENV variables (@mkasztelnik)
 - Possibility to configure elasticsearch URL using ENV variable (@wziajka)
 - Bootstrap JS and font awesome regular (@mszostak)
+- Service tagline (@martaswiatkowska)
 - Transfer tabs from playground to services(@kosmidma)
 - Add styling for services, improve global styles (@kosmidma)
 - User affiliations (@mkasztelnik)